# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's MrSoft News",
        "description": "API's for MrSoft News",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://mrsoft-news.test",
            "description": "Servidor de la API (dinámico: usa APP_URL del entorno actual)"
        }
    ],
    "paths": {
...
        "/api/reel": {
            "get": {
                "tags": [
                    "Reel"
                ],
                "summary": "Get all reels",
                "operationId": "80de2ef32c561052fce0e50a0613ab1c",
                "parameters": [
                    {
                        "name": "product",
                        "in": "query",
                        "description": "Product name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Gesrest",
                                "360sys",
                                "HotelHUB",
                                "Comprobante-e"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of news",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReelResourceCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reel"
                ],
                "summary": "Create a new reel",
                "operationId": "e1fdc4e983d44be22a8d7f023a8b8381",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "url"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Reel title"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://www.youtube.com/watch?v=..."
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReelResourceCollection"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reel/{id}": {
            "get": {
                "tags": [
                    "Reel"
                ],
                "summary": "Get reel by id",
                "operationId": "47671f3045fc171476eed13fc8fa4b73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reel id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReelResourceCollection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reel not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Reel"
                ],
                "summary": "Update a reel",
                "operationId": "b6bed8da7159681e650543f9636539b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reel id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Reel title"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://www.youtube.com/watch?v=..."
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReelResourceCollection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reel not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reel"
                ],
                "summary": "Delete a reel",
                "operationId": "f3c899cfc15d6ae53e079cbbac867e46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reel id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reel deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reel not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/testimonio-tag": {
            "get": {
                "tags": [
                    "Testimonio"
                ],
                "summary": "Get all testimonio tags",
                "operationId": "b02c66c26361c2d2dad274835dd12c77",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of tags",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagTestimonioResourceCollection"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/testimonio-tag/{id}": {
            "get": {
                "tags": [
                    "Testimonio"
                ],
                "summary": "Get testimonio tag by id",
                "operationId": "15687af3bf5779b8ab81691456b647ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagTestimonioResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tag not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "The pagination must be an integer."
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "ReelResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "News title"
                    },
                    "description": {
                        "type": "string",
                        "example": "News description"
                    },
                    "video": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/storage/video.mp4"
                    }
                },
                "type": "object"
            },
            "ReelResourceCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/ReelResource"
                }
            },
            "TagTestimonioResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Facilidad de uso"
                    }
                },
                "type": "object"
            },
            "TagTestimonioResourceCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/TagTestimonioResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}